+++ Web/src/app/inbox/components/InboxContent.tsx
