+++ Web/src/app/layout.tsx
@@ -2,6 +2,7 @@ import type { Metadata } from 'next';
 import { Inter } from 'next/font/google';
 import { Providers } from '@/components/Providers';
 import { Navigation } from '@/components/Navigation';
+import { UsernameSelectionHandler } from '@/features/authentication/components/UsernameSelectionHandler';
 import { ReactNode } from 'react';
 import './globals.css';
 
@@ -28,8 +29,20 @@ export default function RootLayout({ children }: { children: ReactNode }) {
     <html lang="en">
       <body className={`${inter.className}`}>
         <Providers>
-          <Navigation />
-          <main className="min-h-screen">{children}</main>
+          <div className="min-h-screen">
+            {/* Username Selection Handler - must be at root level for modal overlay */}
+            <UsernameSelectionHandler />
+
+            {/* Main Navigation */}
+            <Navigation />
+
+            {/* Main Content */}
+            <main className="flex-1">
+              <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8 py-8">
+                {children}
+              </div>
+            </main>
+          </div>
         </Providers>
       </body>
     </html>
